% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cardiovascular.R
\name{align_drug_start_end_time}
\alias{align_drug_start_end_time}
\title{Align the start and end time for a given drug}
\usage{
align_drug_start_end_time(vasoactive_infusions)
}
\arguments{
\item{vasoactive_infusions}{a df with vasoactive infusions created in \code{\link{get_cv_by_vasoactive_infusion}}}
}
\description{
Align the start and end time for a given drug
}
