% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cardiovascular.R
\name{get_cardiovascular}
\alias{get_cardiovascular}
\title{Joins the age group and vasoactive infusion cardiovascular datasets}
\usage{
get_cardiovascular(cv_by_age_group, cv_by_vasoactive_infusion)
}
\arguments{
\item{cv_by_age_group}{a df returned by \code{\link{get_cv_by_age_group}}}

\item{cv_by_vasoactive_infusion}{a df returned by \code{\link{get_cv_by_vasoactive_infusion}}}
}
\value{
A data frame with the calculated cardiovascular score based on vasoactive infusions
}
\description{
Joins the age group and vasoactive infusion cardiovascular datasets
}
