% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_child_encounter.R
\name{get_child_encounter}
\alias{get_child_encounter}
\title{Transform, filter and expand the child encounter data}
\usage{
get_child_encounter(read_child_encounter)
}
\arguments{
\item{read_child_encounter}{the raw child_encounter csv file obtained from the IDR}
}
\value{
A list containing two data frames named transformed_child_encounter and expanded_child_encounter.
\itemize{
\item Steps 1-3 above were executed on the raw data to create transformed_child_encounter
\item Step 4 was executed on transformed_child_encounter to create expanded_child_encounter
}
}
\description{
This function ingests the raw child encounter data and performs the following:
\itemize{
\item Filters for all subjects less than 22 years old
\item Calculates a unique encounter number for each subject based on the number of occurrences for a given mrn
\item Calculates the length of stay within an encounter
\item Creates the q1hr time buckets
}
}
