% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_coagulation.R
\name{get_coagulation}
\alias{get_coagulation}
\title{Calculate coagulation score}
\usage{
get_coagulation(read_child_labs)
}
\arguments{
\item{read_child_labs}{the raw child labs csv file obtained from the IDR}
}
\value{
A data frame with the calculated coagulation score
}
\description{
Ingests the raw child labs and calculates the coagulation score based on the platelet count.

\itemize{
\item Use the last recorded inferred_specimen_datetime within an hour
\item If there are multiple values for the last recorded inferred_specimen_datetime then use the max coagulation_score
}
}
