% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cardiovascular.R
\name{get_cv_by_age_group}
\alias{get_cv_by_age_group}
\title{Calculate cardiovascular score based on age group}
\usage{
get_cv_by_age_group(read_flowsheets, child_dob)
}
\arguments{
\item{read_flowsheets}{the raw flowsheet data obtained from the IDR}

\item{child_dob}{the dob of each subjects}
}
\value{
A data frame with the calculated cardiovascular score by age group
}
\description{
Calculate cardiovascular score based on age group
}
