% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_respiratory_support.R
\name{get_fio2_value}
\alias{get_fio2_value}
\title{Extract FiO2 Value}
\usage{
get_fio2_value(
  categorized_respiratory_devices,
  read_flowsheets,
  nasal_flow_rate_state
)
}
\arguments{
\item{categorized_respiratory_devices}{the respiratory support devices}

\item{read_flowsheets}{the raw flowsheet csv file obtained from the IDR}

\item{nasal_flow_rate_state}{a df returned by \code{\link{get_nasal_state}}}
}
\value{
A dataframe containing the FiO2 values
}
\description{
Extracts the FiO2 value from the oxygenation and vent settings flowsheets
\itemize{
\item  When there are simultaneously-recorded FiO2 in both the oxygenation and vent settings
flowsheet within an hour chose the oxygenation flowsheet value.

\item When there are more than two FiO2 values within an hour choose the highest value
regardless of if that value comes from  oxygenation or vent settings flowsheet.
}
}
