% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_neurologic.R
\name{get_neurologic}
\alias{get_neurologic}
\title{Calculate neurologic score}
\usage{
get_neurologic(read_glasgow)
}
\arguments{
\item{read_glasgow}{the raw glasgow csv file obtained from the IDR}
}
\value{
A data frame with the neurologic scores
}
\description{
Ingests the glasgow data and calculates the neurologic score.
}
\details{
\itemize{
\item When there are both glasgow_coma_adult_score and glasgow_comma_peds_score recorded for a
given patient on the same encounter the lower score is used to avoid underestimating the severity
\item neurologic scores are carried forward until a new neurologic score is recorded
\item When multiple neurologic scores recorded within an hour the lowest neurologic score is used
}
}
