% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_respiratory_support.R
\name{get_pao2_fio2_resp_score}
\alias{get_pao2_fio2_resp_score}
\title{PaO2/FiO2 Respiratory Score}
\usage{
get_pao2_fio2_resp_score(
  read_child_labs,
  transformed_child_encounter,
  respiratory_support_status
)
}
\arguments{
\item{read_child_labs}{the raw child labs csv file obtained from the IDR}

\item{transformed_child_encounter}{a df returned by \code{\link{get_child_encounter}}}

\item{respiratory_support_status}{a df returned by \code{\link{get_respiratory_status}}}
}
\value{
A dataframe containing the respiratory scores based on the paO2/fiO2 ratio
}
\description{
Calculates the respiratory score based on the PaO2/FiO2 ratio. The PaO2/FiO2 ratio is only
calculated when the PaO2 and FiO2 values are recorded simultaneously
}
