% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_renal.R
\name{get_renal}
\alias{get_renal}
\title{Calculate renal score}
\usage{
get_renal(read_child_labs, child_dob)
}
\arguments{
\item{read_child_labs}{the raw glasgow csv file obtained from the IDR}

\item{child_dob}{the dob of each subjects}
}
\value{
A data frame with the calculated renal score
}
\description{
Ingests the raw lab data data calculates the renal score.
}
\details{
\itemize{
\item Filter for labs that are named "creatine", "creatinine (point of care)",
"creatinine, serum", "creatinine-plasma". Additionally, the lab result must be numeric.
\item When multiple lab results are recorded within an hour the latest lab result is used.
}
}
