% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_respiratory_support.R
\name{get_respiratory_status}
\alias{get_respiratory_status}
\title{Set respiratory status to TRUE/FALSE}
\usage{
get_respiratory_status(read_flowsheets, categorized_respiratory_devices)
}
\arguments{
\item{read_flowsheets}{the raw flowsheets csv file obtained from the IDR}

\item{categorized_respiratory_devices}{an excel file containing the categorized respiratory devices}
}
\value{
A dataframe containing the respiratory status
}
\description{
Sets the respiratory status based on the oxygenation or vent settings device present
\itemize{
\item Instances of mean airway pressure or vent map under vent settings flowsheet indicate on respiratory support
\item Instances of bipap or cpap under oxygenation flowsheet indicate on respiratory support
\item Devices in categorized_respiratory_devices under oxygenation flowsheet indicate not on respiratory support
}
}
