% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Renal.R
\name{Renal}
\alias{Renal}
\title{Renal function that calculates a score based on Creatinine data}
\usage{
Renal(DT, V, E, Age, day_of_interest = 0, clinical_events = c("Creatinine"))
}
\arguments{
\item{DT}{A numeric vector representing event times represented as days from ICU admission (e.g. for an event at 36 hours after admission, DT=1.5)}

\item{V}{A numeric vector representing the results value of Clinical Events associated with DT}

\item{E}{A character vector representing the type of Event assocaited with DT}

\item{Age}{A numeric value representing the age in months}

\item{day_of_interest}{The day for which the renal score should be calculated (non-negative whole number, starts from 0 for the first 24 hours of ICU stay)}

\item{clinical_events}{A character string vector representing the name of the clinical events corresponding to Creatinine (default value is "Creatinine")}
}
\value{
An integer between 0 and 4
}
\description{
Renal function that calculates a score based on Creatinine data
}
