% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Respiratory.R
\name{Respiratory}
\alias{Respiratory}
\title{Respiratory function that calculates a score based on based on Respiratory , SPO2,PaO2, and FiO2 data}
\usage{
Respiratory(
  DT,
  V,
  E,
  day_of_interest = 0,
  maxlag = 6,
  clinical_events = list(Respiratory = c("Respiratory"), SPO2 = c("SPO2"), PaO2 =
    c("PaO2"), FiO2 = c("FiO2"))
)
}
\arguments{
\item{DT}{A numeric vector representing event times represented as days from ICU admission (e.g. for an event at 36 hours after admission, DT=1.5)}

\item{V}{A numeric vector representing the results value of Clinical Events associated with DT}

\item{E}{A character vector representing the type of Event assocaited with DT}

\item{day_of_interest}{The day for which the respiratory score should be calculated (non-negative whole number, starts from 0 for the first 24 hours of ICU stay)}

\item{maxlag}{A positive number for the windwo size in hours to check for FiO2 and  Respiratory support information.}

\item{clinical_events}{A list of 4 string vectors representing the name of the clinical events corresponding to Respiratory, SPO2, PaO2, and FiO2.}
}
\value{
An integer between 0 and 4
}
\description{
Respiratory function that calculates a score based on based on Respiratory , SPO2,PaO2, and FiO2 data
}
